Fix CI failure: MakeGenericType breaks Native AOT -- extract non-generic DxFormBody
CI's AOT-focused analyzer rejected opening a nested DxForm<TNested> via
Type.MakeGenericType (RequiresDynamicCode, incompatible with Native AOT -- this repo
publishes and smoke-tests an actual AOT build in CI).

DxForm<TModel> is now a thin typed wrapper around a new internal, non-generic
DxFormBody, which does the actual field-rendering/validation work entirely over
IFormModelUntyped/object. A nested Object/Array field opens DxFormBody directly --
an ordinary, ahead-of-time-compilable component instantiation, never MakeGenericType.

This also structurally removes the earlier nested-<form>-inside-<form> hazard (fixed
in the previous commit with an internal IsNestedForm flag): DxFormBody renders no
wrapping element at all, so a nested <form> is now impossible by construction, not
just avoided by a flag. The IsNestedForm parameter and IRefreshableForm marker
interface are both gone -- nested refresh propagation now just calls .Refresh()
directly on captured DxFormBody references, since the concrete type is the same
non-generic DxFormBody everywhere.
